TA2549 feat(snap_rebuild) handle snapshot command.
If we are syncing from the cloned replica and we get a snapshot command,
we can not take the snapshot immediately nor we can delay it.
So we have to resync that snapshot from rebuild replica and
then again sync from the clone replica,

We have to handle few more scenario as describe below :-

1. If snapshot command is failed on a healthy replica, it has to disconnect
it so that it can start the rebuilding the process and syncs the snapshot.

2. In AFS mode of replica, when it receives snapshot command,
replica has to disconnect (similar to healthy replica case)

3. In SNAP mode of replica, replica returns error, but,
it will not take any action on data/rebuild connections.

Signed-off-by: Pawan <pawanprakash101@gmail.com>