fix(zfs) double close of the fd and zfs history missing few commands
In case of error we are closing the sock fd immediately, but as part
of error handling, the caller itself closes the fd. So no need to
close the fd and let the caller handle it.

Also few commands are not logged in the zfs history command as they
are not listed as config command in the uzfs IOCTL list, fixed it
by adding them to the uzfs IOCTL list.

Signed-off-by: Pawan <pawanprakash101@gmail.com>