bug(pool_corruption) preventing pool getting imported twice
If we are deleteing a pod, kubrnetes puts this pod into terminating
state and redeploys a new pod. There is a chance that the pool is
imported at two places. Here we are taking a lock on /tmp/zrepl.lock
file to ensure that only one zrepl can start so that only one guy can
import the pool. This file will be there on the persistent storage
so that everybody can see the same file.

Signed-off-by: Pawan <pawanprakash101@gmail.com>