fix(test) fixing incorrect test case
there are two parallel threads trying to write
the data to the volume, one is clone_zv rebuild thread
and other is dw rebuild thread. If the data is overwritten
by other thread, the verification will fail as it verifies
whatever I have written is there on the volume or not.

Fixed it by protecting it by a lock.

Signed-off-by: Pawan <pawanprakash101@gmail.com>